+++ test/test.formatter.js
@@ -284,6 +284,11 @@ describe('formatter', function () {
       helper.assert(stringFormatter.split('coucou', undefined)    , ['coucou']);
       helper.assert(stringFormatter.split('coucou', {})           , ['coucou']);
       helper.assert(stringFormatter.split('coucou', ['ou', 'ao']) , ['coucou']);
+      /** Combine split + arrayJoin */
+      const _resSplit = stringFormatter.split('Lorem ipsum dolor sit amet', ' ');
+      helper.assert(_resSplit, ['Lorem','ipsum','dolor','sit','amet']);
+      const _resJoin = arrayFormatter.arrayJoin(_resSplit, '', '2', '1');
+      helper.assert(_resJoin, 'dolor');
     });
     it('should not crash if data is null or undefined or a regex', function () {
       helper.assert(stringFormatter.split('coucou', /ou/g), ['coucou']);
@@ -1514,10 +1519,17 @@ describe('formatter', function () {
       helper.assert(stringFormatter.substr('coucou', 1, -2), 'ouc');
       helper.assert(stringFormatter.substr('abcdef', -1, 100), 'f');
       helper.assert(stringFormatter.substr('abcdef', -2, 100), 'ef');
+      helper.assert(stringFormatter.substr('coucou', '3', '5'), 'co');
+      helper.assert(stringFormatter.substr('abcdef', '-2', '100'), 'ef');
     });
     it('should not crash if data is null or undefined', function () {
       helper.assert(stringFormatter.substr(null, 0, 3), null);
       helper.assert(stringFormatter.substr(undefined, 0, 3), undefined);
+      helper.assert(stringFormatter.substr([], 0, 3), []);
+      helper.assert(stringFormatter.substr(2, 0, 3), 2);
+      helper.assert(stringFormatter.substr('abcdef', undefined, undefined), '');
+      helper.assert(stringFormatter.substr('abcdef', null, null), '');
+      helper.assert(stringFormatter.substr('abcdef', [], []), '');
     });
     it('should keep only the selection of characters but do not cut words if the third paramater is true. The returned text can be shorter', function () {
       helper.assert(stringFormatter.substr('coucou donotcutme  donotcut   other', 0, 2, true)     , '');
@@ -1530,6 +1542,8 @@ describe('formatter', function () {
       helper.assert(stringFormatter.substr('coucou donotcutme  donotcut   other', 6, 28, true)    , ' donotcutme  donotcut ');
       helper.assert(stringFormatter.substr('coucou donotcutme  donotcut   other', 28, 1000, true) , '  other');
       helper.assert(stringFormatter.substr('coucou donotcutme  donotcut   other', 28, 35, true)   , '  other');
+      helper.assert(stringFormatter.substr('coucou donotcutme  donotcut   other', '0', '20', 'true')  , 'coucou donotcutme  ');
+      helper.assert(stringFormatter.substr('coucou donotcutme  donotcut   other', '6', '28', true)    , ' donotcutme  donotcut ');
 
       helper.assert(stringFormatter.substr('coucou donotcutme  donotcut   other', 0, -2, true)    , 'coucou donotcutme  donotcut   ');
       helper.assert(stringFormatter.substr('coucou donotcutme  donotcut   other', 0, -1, true)    , 'coucou donotcutme  donotcut   ');
@@ -1710,6 +1724,11 @@ describe('formatter', function () {
       helper.assert(arrayFormatter.arrayJoin(['1', '2', '3', '4'], '', 3), '4');
       helper.assert(arrayFormatter.arrayJoin(['1', '2', '3', '4'], '', 4), '');
       helper.assert(arrayFormatter.arrayJoin(['1', '2', '3', '4'], '', 5), '');
+      helper.assert(arrayFormatter.arrayJoin(['1', '2', '3', '4'], '', '1', '0'), '');
+      helper.assert(arrayFormatter.arrayJoin(['1', '2', '3', '4'], '', '1'), '234');
+      helper.assert(arrayFormatter.arrayJoin(['1', '2', '3', '4'], '', '1', '1'), '2');
+      helper.assert(arrayFormatter.arrayJoin(['1', '2', '3', '4'], '', '1', '2'), '23');
+      helper.assert(arrayFormatter.arrayJoin(['1', '2', '3', '4'], '', '1', '100'), '234');
     });
     it('should not crash if datas is null or undefined', function () {
       helper.assert(arrayFormatter.arrayJoin(null), null);
@@ -1717,6 +1736,11 @@ describe('formatter', function () {
       helper.assert(arrayFormatter.arrayJoin(120), 120);
       helper.assert(arrayFormatter.arrayJoin([]), '');
       helper.assert(arrayFormatter.arrayJoin({}), {});
+      helper.assert(arrayFormatter.arrayJoin(['1', '2', '3', '4'], '', null, null), '');
+      helper.assert(arrayFormatter.arrayJoin(['1', '2', '3', '4'], '', {}, {}), '');
+      helper.assert(arrayFormatter.arrayJoin(['1', '2', '3', '4'], '', [], []), '');
+      helper.assert(arrayFormatter.arrayJoin(['1', '2', '3', '4'], '', 'null', 'null'), '');
+      helper.assert(arrayFormatter.arrayJoin(['1', '2', '3', '4'], '', -2, -2), '');
     });
   });