+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### v4.12.2
+  - Fix chained formatters `:split():arrayJoin('', 2, 3)`
+  - Fix formatter `substr`
+
 ### v4.12.1
   - Release June 28th 2023
   - Fix `ellispis` -> `ellipsis` formatter